asynchronous generation of keys

+++ deploy.sh
@@ -1,25 +1,29 @@
 #!/bin/sh
 
-SW_PATH="/home/node-7/visible/platforms/node"
+if [ "$(id -u)" != "0" ]; then
+   echo "This script must be run as root" 1>&2
+   exit 1
+fi
 
-cd $SW_PATH;
+SW_PATH=$(pwd);
+
 git stash;
 git fetch;
 git reset --hard origin/master;
-cd $SW_PATH/lib;
-sleep 1;
-cp Config.js Config.js.org ;
-cp Config.js.bkp Config.js;
-sleep 1;
-su - node-7 -c "forever stopall"
-sleep 3;
-cd $SW_PATH;
-su - node-7 -c "forever start -l $SW_PATH/instance_0.log -a $SW_PATH/server.js --instanceNumber=0 & "
-su - node-7 -c "forever start -l $SW_PATH/instance_1.log -a $SW_PATH/server.js --instanceNumber=1 & "
-su - node-7 -c "forever start -l $SW_PATH/instance_2.log -a $SW_PATH/server.js --instanceNumber=2 & "
 
-su - node-7 -c "forever start -l $SW_PATH/instance_3.log -a $SW_PATH/server.js --instanceNumber=3 & "
-su - node-7 -c "forever start -l $SW_PATH/instance_4.log -a $SW_PATH/server.js --instanceNumber=4 & "
 
+forever stopall
 sleep 3;
-su - node-7 -c "tail -f $SW_PATH/instance_?.log "
+
+targetCnt=`cat $SW_PATH/server/lib/installer/NUMBER_OF_INSTANCES.dat`
+NUMBER_OF_INSTANCES=$(($targetCnt + 0))
+COUNTER=0
+while [  $COUNTER -lt $NUMBER_OF_INSTANCES ]; do
+  forever start -l $SW_PATH/server/log/instance_$COUNTER.log -a $SW_PATH/server/server.js --instanceNumber=$COUNTER &
+  let COUNTER=COUNTER+1 
+done
+sleep 3;
+
+tail -f $SW_PATH/server/log/instance_?.log
+
+
